+++ packages/js/product-editor/src/blocks/inventory-sku/index.ts
@@ -1,11 +1,16 @@
 /**
+ * External dependencies
+ */
+import { BlockConfiguration } from '@wordpress/blocks';
+
+/**
  * Internal dependencies
  */
-import initBlock from '../../utils/init-block';
-import metadata from './block.json';
+import { initBlock } from '../../utils/init-block';
+import blockConfiguration from './block.json';
 import { Edit } from './edit';
 
-const { name } = metadata;
+const { name, ...metadata } = blockConfiguration as BlockConfiguration;
 
 export { metadata, name };
