+++ packages/js/product-editor/src/utils/index.ts
@@ -22,7 +22,7 @@ import { isValidEmail } from './validate-email';
 
 export * from './create-ordered-children';
 export * from './sort-fills-by-order';
-export * from './init-blocks';
+export * from './init-block';
 export * from './product-apifetch-middleware';
 export * from './sift';
