--- packages/js/product-editor/src/utils/init-blocks.ts
@@ -1,31 +1,0 @@
-/**
- * External dependencies
- */
-import {
-	Block,
-	BlockConfiguration,
-	registerBlockType,
-} from '@wordpress/blocks';
-
-interface BlockRepresentation< T extends Record< string, object > > {
-	name?: string;
-	metadata: BlockConfiguration< T >;
-	settings: Partial< BlockConfiguration< T > >;
-}
-
-/**
- * Function to register an individual block.
- *
- * @param  block The block to be registered.
- * @return The block, if it has been successfully registered; otherwise `undefined`.
- */
-export function initBlock<
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	T extends Record< string, any > = Record< string, any >
->( block: BlockRepresentation< T > ): Block< T > | undefined {
-	if ( ! block ) {
-		return;
-	}
-	const { metadata, settings, name } = block;
-	return registerBlockType< T >( { name, ...metadata }, settings );
-}
